Refactor BookmarksHostingCell
