+++ podcasts/PodcastViewController+TableData.swift
@@ -4,31 +4,6 @@ import UIKit
 import PocketCastsServer
 import SwiftUI
 
-class TransparentBookmarksStyle: ThemedBookmarksStyle {
-    override var background: Color { Color.clear }
-}
-
-class BookmarksHostingCell: UITableViewCell {
-    static let reuseIdentifier = "BookmarksListCell"
-
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        selectionStyle = .none
-        backgroundColor = .clear
-        contentView.backgroundColor = .clear
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    // Prevent this cell from visually entering editing mode (no indentation or controls)
-    override func setEditing(_ editing: Bool, animated: Bool) {
-        // Intentionally ignore editing state to avoid shifting/indenting this embedded SwiftUI view
-        super.setEditing(false, animated: animated)
-    }
-}
-
 extension PodcastViewController: UITableViewDataSource, UITableViewDelegate {
     private static let episodeCellId = "EpisodeCell"
     private static let headerCellId = "HeaderCell"
@@ -243,30 +218,9 @@ extension PodcastViewController: UITableViewDataSource, UITableViewDelegate {
                     return UITableViewCell()
                 }
                 let cell = tableView.dequeueReusableCell(withIdentifier: BookmarksHostingCell.reuseIdentifier, for: indexPath) as! BookmarksHostingCell
-                cell.backgroundColor = .clear
-                cell.contentView.backgroundColor = .clear
-
-                // Wrap in a top-aligned container to avoid temporary vertical centering while sizing updates.
-                cell.contentConfiguration = UIHostingConfiguration {
-                    VStack(spacing: 0) {
-                        BookmarksListView(viewModel: bookmarkViewModel,
-                                          style: TransparentBookmarksStyle(),
-                                          showHeader: true,
-                                          showMultiSelectInHeader: false,
-                                          allowInternalScrolling: false,
-                                          showSearchField: true,
-                                          useExternalActionBar: true,
-                                          externalActionBarHandler: { [weak self] state in
-                            self?.updateBookmarksActionBar(state: state, viewModel: bookmarkViewModel)
-                        })
-                        .padding(.top, 10)
-
-                        // Fills remaining space so content stays pinned to the top.
-                        Spacer(minLength: 0)
-                    }
-                    .frame(maxWidth: .infinity, alignment: .topLeading)
+                cell.configure(with: bookmarkViewModel) { [weak self] state in
+                    self?.updateBookmarksActionBar(state: state, viewModel: bookmarkViewModel)
                 }
-                .margins(.all, 0)
                 return cell
             }
